refactor: change the navigation to the new search
Remove references to the old search fragment and point to the new search instead
