--- app-discover/src/main/res/menu/menu_discover.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:showIn="@layout/fragment_discover">
-
-    <item
-        android:id="@+id/menuDiscoverSearch"
-        android:icon="@drawable/ic_search"
-        android:title="@string/discover_search"
-        app:actionViewClass="androidx.appcompat.widget.SearchView"
-        app:showAsAction="ifRoom" />
-</menu>
