+++ app-search/src/main/java/com/chesire/malime/app/search/SearchViewModel.kt
@@ -1,0 +1,7 @@
+package com.chesire.malime.app.search
+
+import androidx.lifecycle.ViewModel
+import javax.inject.Inject
+
+class SearchViewModel @Inject constructor() : ViewModel() {
+}
