+++ app/src/main/java/com/chesire/malime/injection/androidmodules/FragmentModule.kt
@@ -1,8 +1,8 @@
 package com.chesire.malime.injection.androidmodules
 
 import com.chesire.malime.app.discover.DiscoverFragment
-import com.chesire.malime.app.discover.search.SearchFragment
 import com.chesire.malime.app.profile.ProfileFragment
+import com.chesire.malime.app.search.SearchFragment
 import com.chesire.malime.app.series.detail.SeriesDetailSheetFragment
 import com.chesire.malime.app.series.list.anime.AnimeFragment
 import com.chesire.malime.app.series.list.manga.MangaFragment
@@ -34,12 +34,12 @@ abstract class FragmentModule {
     @ContributesAndroidInjector
     abstract fun contributeProfileFragment(): ProfileFragment
 
+    @ContributesAndroidInjector
+    abstract fun contributeSearchFragment(): SearchFragment
+
     @Suppress("FunctionMaxLength")
     @ContributesAndroidInjector
     abstract fun contributeSeriesDetailSheetFragment(): SeriesDetailSheetFragment
-
-    @ContributesAndroidInjector
-    abstract fun contributeSeriesSearchFragment(): SearchFragment
 
     @ContributesAndroidInjector
     abstract fun contributeSyncingFragment(): SyncingFragment
